Restructure pytesting: conftest, test_IB and test_UC
